refactor: move shape module headers to include/core/
Reorganize project structure so all public headers live in include/core/
and implementation stays in src/. Updated all #include paths accordingly.

Before:
  src/*.h  (scattered headers)

After:
  include/core/*.h  (all headers centralized)
  src/*.c  (implementations only)
