+++ include/core/shape_geometry.h
