+++ include/core/shape_impl.h
