+++ include/core/shape_manager.h
